Fixed a few things: movement_check() was being called after every move, should be only once per step. on_land() wasn't properly delegating to in_water(). ['p_long'] and ['p_lat'] were transposed. c_gnome.wind_mover wasn't moving particles because their z-positions hadn't been initialized and so retained a garbage value. changed world_point type now to represent the lats and longs as floats instead of long longs, and so modified only slightly the wrapped mover functions. some other minor things, here and there.

+++ py_gnome/cygnome/codeFiles/TypeDefs.h
@@ -310,8 +310,8 @@ typedef struct {
 
 
 typedef struct {
-	long pLong;
-	long pLat;
+	float pLong;
+	float pLat;
 } WorldPoint, *WORLDPOINTP, **WORLDPOINTH;
 
 typedef struct {
